Move to grub2, boot only from virtual disk image.
Switched to grub2 and wrote mkdisk.sh to autogenerate the raw disk
image. It is very easy now to import this image to virtualbox if you want
(see the end of mkdisk.sh).
When type "make run", now it will generate the hdd.img disk, add 2 partitions,
one for boot and one for COFS formated disk, copy all the binaries to it.
It will install grub2 on the MBR of the virtual disk image, mount it,
copy all binaries to it and fire up bochs.
Right now partitions are hardcoded to fixed sectors; a partition
autodetection algorithm must be written.