Make the serial line to act like a console
The serial (COM1) was used only to dump text to serial port (and wrote
by bochs to serial.out). I decided it will be nice to make it act
more like a console, so i cleaned up, moved the 'make run' to start qemu
(apt-get install qemu-system-x86) with stdio attached to it and no display.
It boots faster this way. You can still run it in bochs with 'make runb'.
Idea is that on real hardware, with no display and keyboard, you can
use a serial null modem (cross cable) and still use the OS remotely.
I also plan to use COM2 for general early (when there is no irq
available) debug, or better for remote debug with gdb.
Files console.c and kbd.c needs to be reworked, because it becomes messy.