Fix init to rip zombies, clean scheduler
Init should go to sleep after forking the shell (cosh) and wakeup when a
zombie process dies. If cosh dies, it should respawn it. Also, the
scheduler will wakeup init when it has no other process to switch; init
should ignore wait() == -1 return. It is hacky a little bit but i don't
want to wake other processes that sleeps in critical zones.
There is no need for cli instruction in interrupt handlers,
because i386 disable interrupts by default when entering. We will permit
isr128 (0x80, syscall) to be interrupted, because a process that just
switched from ring 3 to 0 can go to sleep.
Because some of the functions, like fork() and task_switch() can
be called from syscall, we will ensure the interrupts are off in this
particular (and critical) functions.