Use only one task and stacks to handle signals
A better approach on handling custom signal handlers is to use same
process, (and not a new thread as in previous commit) but with different kernel
stack. When starting to handle a signal, the current values of the task are
saved in a pointer (s_task). It's kernel stack is moved down, the task selector
is informed of it, so that on the next switch from ring 3 to ring 0, this new
low portion of the stack will be used, leaving the current kernel stack space
in it's place.
Another idea is to push the current task on the kernel stack, but, it's
easier to use a new allocated memory.
On returning from signal handler, the original task will be restored and
stack will be moved again to KERNEL_STACK_HI.
Added default signals.