Write a vsnprintf lib function
A more secure vsnprintf function is needed in the kernel.
Rewrite all other function that made use of unsafe function vsprintf to
use this new one. It has some limitations - right now it does not
provide floating points numbers and there is a known bug on ll, L length
modifiers, because we don't have yet a div64 function to use in fmt_int,
and I don't want to complicate the things too much. That's it, it will
print only 32 bits numbers for now.