Switch to ELF for user programs
We should use Executable and Linkable Format for user/ring3 programs,
because it is easy to debug and it's more standard.
Kernel will use ELF in sys_exec() from now on (see sysproc.c).
Loader is still very basic, more tests can be done - check for user
permissions in the future when it will be multiuser, check for
architecture, etc. Programs can be loaded automatically, as long as they
are compiled to fit inside USER_CODE_START_ADDR - UHEAP_START (see
README.txt).